requirements.txt
tensorflow==2.15.0
numpy
Pillow
opencv-python
streamlit
matplotlib
pandas
plotly
keras